Move addTicketUrls to separate mutation, refactor ticket queries
